Crate sorella_db_tracing

clickhouse tracing subscribers and utilities.

Contains a standardized set of layers:

stdout()

file()

journald()

As well as a simple way to initialize a subscriber: init.

Re-exports

pub use tracing;

pub use tracing_subscriber;

Modules

colors

colors module

fmt

formating tracing with colors

Macros

log

logs a message of strings and values color formats the values

Functions

file

Builds a new tracing layer that appends to a log file.

init

Initializes a new Subscriber based on the given layers.

journald

Builds a new tracing layer that writes events to journald.

stdout

Builds a new tracing layer that writes to stdout.

testing_default

default for testing

Type Aliases

BoxedLayer

A boxed tracing Layer.

FileWorkerGuard

A worker guard returned by file().